+++ CHANGELOG.md
@@ -6,3 +6,4 @@ All notable changes to this project will be documented in this file.
 
 - Add initial project structure
 - Add creator vertical ([#1](https://github.com/nahuelsoma/event-driven-challenge-payments/pull/1))
+- Add finder vertical ([#2](https://github.com/nahuelsoma/event-driven-challenge-payments/pull/2))